Dietary preferences and restrictions multi-select
The profile page now contains a multi-select option for the dietary restrictions and dietary preferences when pressing the "Edit" button. The profile also no longer has an "Edit all" option so that there is no redundancy. Will have to update previous task cards to reflect this as well. Daily Health Tips is now available at the bottom of the profile page and contains a variety of different health tips.