profile contains most functionality
This branch contains only the necessary files for the profile page. It has all the functionality minus the daily health tips and the dietary restrictions and dietary preferences being multi-select. This version is an earlier version of the branch 80-Multi-Select-Dietary-Preferences. The only files that remain unchanged from this branch to that branch are update_profile.php and logout.php.