Change dependencies and redirection mechanism
In the previous version of registration.php, registration.php would not redirect to login.php. This version aims to resolve that issue by resolving dependencies on vendor and checks if it exists first. It also resolves the redirection mechanism for specifically the development server by checking if the absolute path exists first and if not, will redirect to the appropriate path: login.php.