Half functional dashboard
Dashboard can now be accessed from the profile by pressing on the "Dashboard" button in profile. This will redirect the user to the dashboard page where the weekly meal plan can be set. The bookmark function is yet to be implemented, so for now, pressing on the green + button will open a modal to the search page where the database of recipes will be displayed.

Improvements to be made (?):
- Modal shows bookmarks for that day only
- Modal shows favorited bookmarks
- Modal allows users to bookmark a recipe from the page itself