Added header to all logged in pages
This commit implements the header for all pages where the user is logged in. To format the header properly on the pages, a few lines of HTML were removed. On top of that, map.html and about.html are now PHP files so that the header can function properly on these pages. Minimal code changes were made to existing files only to implement the headers and not change the pages' functionalities. This may result in database errors locally but **may** not affect development. This commit should also update the entire meal plan at once rather than using separate actions in dashboard.php.